refactor(job_scrapper): refactor sql component to ./sql
