+++ job_scrapper/scrapper_skeleton/request_core.py
@@ -23,7 +23,7 @@ from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.support import expected_conditions as EC
 from selenium.webdriver.support.ui import WebDriverWait
 
-from .sql_core.sql_core import ScrapperSQLightCore
+from .sql_core import ScrapperSQLightCore
 
 
 class ScrapperRequestCore(ScrapperSQLightCore):
