+++ job_scrapper/scrapper_skeleton/sql_core.py
@@ -9,11 +9,11 @@ from sqlalchemy.orm import Session, Query
 from sqlalchemy import case, func, and_, or_, not_, Result
 
 
-from ..object_core import ScrapperObjectCore
-from .sql_tables import BaseTableForJobScrapper, Jobs, Metadata, TimeStamps, Distances, Keywords
-from .filter_generator import FilterPart, FilterGenerator
-from .wrapper_comparison import STRING_TO_COMPARISON_WRAPPERS
-from .wrapper_logical import STRING_TO_LOGICAL_WRAPPERS
+from job_scrapper.scrapper_skeleton.object_core import ScrapperObjectCore
+from sql.tables.sql_tables import BaseTableForJobScrapper, Jobs, Metadata, TimeStamps, Distances, Keywords
+from sql.filters.filter_generator import FilterPart, FilterGenerator
+from sql.wrappers.wrapper_comparison import STRING_TO_COMPARISON_WRAPPERS
+from sql.wrappers.wrapper_logical import STRING_TO_LOGICAL_WRAPPERS
 from sqlalchemy.sql import operators as ope
 
 class ScrapperSQLightCore(ScrapperObjectCore):
